% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_provisioning_artifact}
\alias{servicecatalog_describe_provisioning_artifact}
\title{Gets information about the specified provisioning artifact (also known
as a version) for the specified product}
\usage{
servicecatalog_describe_provisioning_artifact(AcceptLanguage,
  ProvisioningArtifactId, ProductId, Verbose)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{ProvisioningArtifactId}{[required] The identifier of the provisioning artifact.}

\item{ProductId}{[required] The product identifier.}

\item{Verbose}{Indicates whether a verbose level of detail is enabled.}
}
\description{
Gets information about the specified provisioning artifact (also known
as a version) for the specified product.
}
\section{Request syntax}{
\preformatted{servicecatalog$describe_provisioning_artifact(
  AcceptLanguage = "string",
  ProvisioningArtifactId = "string",
  ProductId = "string",
  Verbose = TRUE|FALSE
)
}
}

\keyword{internal}
