% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_list_principals_for_portfolio}
\alias{servicecatalog_list_principals_for_portfolio}
\title{Lists all principal ARNs associated with the specified portfolio}
\usage{
servicecatalog_list_principals_for_portfolio(AcceptLanguage,
  PortfolioId, PageSize, PageToken)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{PortfolioId}{[required] The portfolio identifier.}

\item{PageSize}{The maximum number of items to return with this call.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of
results, use null.}
}
\description{
Lists all principal ARNs associated with the specified portfolio.
}
\section{Request syntax}{
\preformatted{servicecatalog$list_principals_for_portfolio(
  AcceptLanguage = "string",
  PortfolioId = "string",
  PageSize = 123,
  PageToken = "string"
)
}
}

\keyword{internal}
